requirements.txt
# AI Employee - All Required Dependencies
# Install with: pip install -r requirements.txt

# Core
watchdog>=4.0.0
python-dotenv>=1.0.0

# Google/Gmail API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.23.0

# Web Automation
playwright>=1.40.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# Social Media APIs
tweepy>=4.14.0

# Odoo Integration
odoo-addon-api>=1.0.0

# System Monitoring
psutil>=5.9.0

# Async Support
asyncio>=3.4.3

# Data Processing
pyyaml>=6.0.1

# File Watching
asyncwatch>=0.1.0

# Error Recovery
tenacity>=8.2.3

# Logging
colorlog>=6.8.0
